refactor: introduce use-multiselect
